Refactor to split out get_contributors_for_repo function
