+++ src/__init__.py
@@ -1,4 +1,5 @@
 from src.make_data.find_organisation_repos import find_organisation_repos
-from src.make_data.get_contributors_for_all_repos import get_contributors_for_all_repos, get_contributors_for_repo
+from src.make_data.get_contributors_for_all_repos import get_contributors_for_all_repos
+from src.make_data.get_contributors_for_repo import get_contributors_for_repo
 from src.utils.logger import Log, logger
 from src.utils.parse_api_token import parse_api_token
